refactor(search-sql): extract CompilationDiagnosticsBuilder from SearchCompiler
Copilot-Session: 75061886-abb9-46f7-89dd-a1e40f3d4fff
